test: split gemini preset coverage
